Move editable table to forms-utils.jsx
